requirements.txt
# 基础依赖
python-docx>=0.8.11
python-pptx>=0.6.21
openpyxl>=3.0.10
openai>=0.27.0
PyQt5>=5.15.0
python-dotenv>=0.19.0

# 打包工具
pyinstaller>=5.6.2